Move tcp and grpc probe tests to common files
